Use meson build
Having a build system helps distributors, and allows develpers to easily use
various compiler options and other testing facilities such as sanitizes.  As
small additional thing, the 'make check' will now run couple config tests.

Reference: http://mesonbuild.com/
