--- src/version.h
@@ -1,39 +1,0 @@
-/*
- * Intel(R) Enclosure LED Utilities
- * Copyright (C) 2009-2018 Intel Corporation.
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms and conditions of the GNU General Public License,
- * version 2, as published by the Free Software Foundation.
- *
- * This program is distributed in the hope it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin St - Fifth Floor, Boston, MA 02110-1301 USA.
- *
- */
-
-#ifndef _VERSION_H_INCLUDED_
-#define _VERSION_H_INCLUDED_
-
-/**
- * The major number of LEDCTL utilities package. This number may be overwrite
- * in makefile to reflect current version number.
- */
-#ifndef VERSION_MAJOR
-#define VERSION_MAJOR    0
-#endif
-
-/**
- * The minor number of LEDCTL utilities package. This number may be overwrite
- * in makefile to reflect current version number.
- */
-#ifndef VERSION_MINOR
-#define VERSION_MINOR    90
-#endif
-
-#endif				/* _VERSION_H_INCLUDED_ */
