+++ tests/expected/empty_file
@@ -1,0 +1,9 @@
+INTERVAL: 0
+LOG_LEVEL: 0
+LOG_PATH: (null)
+BLINK_ON_MIGR: 0
+BLINK_ON_INIT: 0
+REBUILD_BLINK_ON_ALL: 0
+RAID_MEMBERS_ONLY: 0
+WHITELIST: NONE
+BLACKLIST: NONE
