+++ tests/expected/invalid_options
@@ -1,0 +1,2 @@
+config file: unknown option 'INTERVAL: 15'.
+invalid_options: parse error
