+++ tests/expected/valid_options
@@ -1,0 +1,9 @@
+INTERVAL: 15
+LOG_LEVEL: 5
+LOG_PATH: /tmp/foo
+BLINK_ON_MIGR: 0
+BLINK_ON_INIT: 1
+REBUILD_BLINK_ON_ALL: 1
+RAID_MEMBERS_ONLY: 0
+WHITELIST: /sys/devices/foo, 
+BLACKLIST: /sys/devices/bar,
